requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
flask-heroku==0.1.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
oauthlib==3.1.1
pycoingecko==2.2.0
python-dotenv==0.19.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
SQLAlchemy==1.4.26
tweepy==4.1.0
urllib3==1.26.7
Werkzeug==2.0.2
